requirements.txt
allure-pytest
appium-python-client
pytest
selene==2.0.0b16